[ etc ] 스타일 관련 오류 수정
