+++ styles/GlobalStyle.tsx
@@ -21,7 +21,6 @@ const GlobalStyle = createGlobalStyle`
     body {
         width: 100%;
         height: 100vh;
-        height: calc(var(--vh, 1vh) * 100);
         padding: 0;
         margin: 0;
         overflow-x: hidden;
